[38] Fix some warnings & typos

+++ src/LorrisAnalyzer/packet.cpp
@@ -62,21 +62,21 @@ bool analyzer_data::getCmd(quint8 &cmd)
 
 quint8 analyzer_data::getUInt8(quint32 pos)
 {
-    if(pos >= m_data.length())
+    if(pos >= (quint32)m_data.length())
         return 0;
     return (quint8)m_data[pos];
 }
 
 qint8 analyzer_data::getInt8(quint32 pos)
 {
-    if(pos >= m_data.length())
+    if(pos >= (quint32)m_data.length())
         return 0;
     return (qint8)m_data[pos];
 }
 
 quint16 analyzer_data::getUInt16(quint32 pos)
 {
-    if(pos+2 > m_data.length())
+    if(pos+2 > (quint32)m_data.length())
         return 0;
     quint16 val = ((quint8)m_data[pos] << 8) | (quint8)m_data[pos+1];
     if(!m_packet->big_endian)
@@ -86,7 +86,7 @@ quint16 analyzer_data::getUInt16(quint32 pos)
 
 qint16 analyzer_data::getInt16(quint32 pos)
 {
-    if(pos+2 > m_data.length())
+    if(pos+2 > (quint32)m_data.length())
         return 0;
     qint16 val = ((quint8)m_data[pos] << 8) | (quint8)m_data[pos+1];
     if(!m_packet->big_endian)
@@ -96,7 +96,7 @@ qint16 analyzer_data::getInt16(quint32 pos)
 
 quint32 analyzer_data::getUInt32(quint32 pos)
 {
-    if(pos+4 > m_data.length())
+    if(pos+4 > (quint32)m_data.length())
         return 0;
     quint32 val = ((quint8)m_data[pos++] << 24) | ((quint8)m_data[pos++] << 16) |
                   ((quint8)m_data[pos++] << 8)  | (quint8)m_data[pos];
@@ -107,7 +107,7 @@ quint32 analyzer_data::getUInt32(quint32 pos)
 
 qint32 analyzer_data::getInt32(quint32 pos)
 {
-    if(pos+4 > m_data.length())
+    if(pos+4 > (quint32)m_data.length())
         return 0;
     qint32 val = ((quint8)m_data[pos++] << 24) | ((quint8)m_data[pos++] << 16) |
                   ((quint8)m_data[pos++] << 8)  | (quint8)m_data[pos];
@@ -118,12 +118,12 @@ qint32 analyzer_data::getInt32(quint32 pos)
 
 quint64 analyzer_data::getUInt64(quint32 pos)
 {
-    if(pos+8 > m_data.length())
+    if(pos+8 > (quint32)m_data.length())
         return 0;
-    quint64 val = ((quint8)m_data[pos++] << 56) | ((quint8)m_data[pos++] << 48) |
-                  ((quint8)m_data[pos++] << 40) | ((quint8)m_data[pos++] << 32) |
-                  ((quint8)m_data[pos++] << 24) | ((quint8)m_data[pos++] << 16) |
-                  ((quint8)m_data[pos++] << 8)  | (quint8)m_data[pos];
+    quint64 val = ((quint64)m_data[pos++] << 56) | ((quint64)m_data[pos++] << 48) |
+                  ((quint64)m_data[pos++] << 40) | ((quint64)m_data[pos++] << 32) |
+                  ((quint32)m_data[pos++] << 24) | ((quint32)m_data[pos++] << 16) |
+                  ((quint16)m_data[pos++] << 8)  | (quint8)m_data[pos];
     if(!m_packet->big_endian)
         val = Utils::swapEndian64(val);
     return val;
@@ -131,12 +131,12 @@ quint64 analyzer_data::getUInt64(quint32 pos)
 
 qint64 analyzer_data::getInt64(quint32 pos)
 {
-    if(pos+8 > m_data.length())
+    if(pos+8 > (quint32)m_data.length())
         return 0;
-    qint64 val = ((quint8)m_data[pos++] << 56) | ((quint8)m_data[pos++] << 48) |
-                 ((quint8)m_data[pos++] << 40) | ((quint8)m_data[pos++] << 32) |
-                 ((quint8)m_data[pos++] << 24) | ((quint8)m_data[pos++] << 16) |
-                 ((quint8)m_data[pos++] << 8)  | (quint8)m_data[pos];
+    qint64 val = ((quint64)m_data[pos++] << 56) | ((quint64)m_data[pos++] << 48) |
+                 ((quint64)m_data[pos++] << 40) | ((quint64)m_data[pos++] << 32) |
+                 ((quint32)m_data[pos++] << 24) | ((quint32)m_data[pos++] << 16) |
+                 ((quint16)m_data[pos++] << 8)  | (quint8)m_data[pos];
     if(!m_packet->big_endian)
         val = Utils::swapEndian64(val);
     return val;
@@ -145,9 +145,9 @@ qint64 analyzer_data::getInt64(quint32 pos)
 QString analyzer_data::getString(quint32 pos)
 {
     QString str = "";
-    if(pos >= m_data.length())
+    if(pos >= (quint32)m_data.length())
         return str;
-    for(; pos < m_data.length() && QChar(m_data[pos]) != 0; ++pos)
+    for(; pos < (quint32)m_data.length() && QChar(m_data[pos]) != 0; ++pos)
         str.append(QChar(m_data[pos]));
     return str;
 }
